Remove unneeded question when certain conditions are met
For employed and self-employed mothers who have not worked
at least 26 weeks, their outcome can already be determined
with this knowledge.

Previously we have directed them to answer the next question
on earnings, but their input here made no difference to their
forward journey or outcomes.

This commit adapts the code from the earnings question,
creating some duplication in the flow but leading to a
better user experience.

Users who do not meet these conditions will continue to be
asked the earnings question, and from directed on the same
path.