Add in details component to smart answers
This will be needed as part of the onboarding part of our banner.

Had to set a fixed with to the first column as that was the only way to ensure the layout was set correctly without creating a custom grid.

Tried using flex and CSS grids but it seems to overcomplicate the solution and as it's not widely used on GOVUK, it was decided that it would be better without this.

It is important to note that the width set will only work for the text that we currently have, as the text is static, this isn't a problem right now.
If we do in the future decide to change the text, then we should consider using flex or css grids to implement a custom layout.