Simplify the configuration of hide-this-page
There doesn't seem to be a need to have two similar questions about
whether to hide a page or not, so I have combined them.

There is an edge case that that a Smart Answer cannot be configured with
a `use_hide_this_page` before a `response_store` (otherwise you'll
receive an exception) but this seems a minor concern given how few flows
utilise this feature (1).