Iterate State to better support response_store flows
This changes the state object to have a hash, accepted_responses, which
stores a mapping of node name to answer for any answered questions.
This replaces previous values of path (an array of node names) and
response (an array of answers). This has been done for better suitability
with the resonse_store flows which deal with hashes of responses rather
than arrays.

It renames the response field to be current_response to better
distinguish it from accepted_responses (the previous approach of
responses and response was very confusing for me).

It adds in a constructor parameter of forwarding responses which can be
used as a mechanism to record any responses that can be passed with all
answers. Unlike accepted_responses, which only stores valid responses to
questions earlier in the flow, forwarding_responses will hold every
response a user has given in a flow. It's intention is for the usage in
any URL's to help navigate around a Smart Answer to allow storing past
answers.

As part of this I removed some tests that didn't seem necessary, as they
were basically testing OpenStruct.