Add #has? to ResponseStore
This method is added to allow checking whether a parameter is stored in
a session response store it will be used to determine whether an answer
has been given by a user even if it was empty. It's needed because the #get
method will return nil for non-existent answers and empty answers.