Utilise State to simplify FlowPresenter
Now that the State object better represents a users current state in
a flow, various code can be removed from FlowPresenter that instead
makes use of the State object.

Rather than accepting a param object and creating a state this now
accepts a state object directly. This creates a degree of abstraction
which will mean that parameters no longer need to be coerced into
meeting this interface [1] and remove the need to consider whether
ActionController::Parameters are part of the input.

These changes here simplify some of the ideas that need to be understood
by the class where there were previously a whole variety of different
ways to access responses and different names (all_responses,
accepted_responses). Instead all previous responses are accessed via the
state object.

This also removes the param attribute from NodePresenters as it is no
longer available in a FlowPresenter and doesn't seem to be used at all
in these classes anyway.

[1]: https://github.com/alphagov/smart-answers/blob/2bba0fba144240641577a88e6e88bd537eb133b6/app/helpers/flow_helper.rb#L8