Update FlowController and FlowHelper for updated FlowPresenter
This class now creates state objects before populating the FlowPresenter
object. This is not done in a shared method as there is subtle
distinction between what is needed for #update and #show. In #update we
want to always take the user to the next available step in the flow,
whereas in the #show action we want the request to specify which node is
wanted. State needs to take this into consideration.

As some of the FlowHelper methods are now covered by state these can now
be removed.