Port prefilling answers to using state.current_response
This replaces the use of the #response_for_current_question method to
instead utilise the #current_response method that is provided by the
State object.

A complication we have with the #current_response object on State is that
the value could be whatever object was initially input (string, array,
hash) or a string (from a previous_response or path parameter) thus any
code that has to establish a previous value should consider the object
could be any type. Although this seems like a complication it seems
beneficial to me as type concerns existed before but were abstracted
away.

This removes the `to_response` class that was provided on various
QuestionPresenters before in favour of the individual checking.
This #to_response was a bit confusing as it was intended for converting
a string (params[:previous_response]) into a parsed object, however since
the introduction of FlowController past responses aren't necessarily
string representations.

I removed some utility methods in the Salary and Date classes to instead
return hashes. I did this to make testing simpler since I couldn't think
of an elegant way to test them all individually and succinctly. I don't
think we lose much readability by this change.