Remove state code from Flow
This handling is now done by the ResolveState class which allows the
code here (and the substantial quantity of tests) to be removed.

There was a bunch of tests that broke in this because the FlowTestHelper
does a rather clunky faking of a responses parameter. I fixed these by
changing all the dates to be conventional formatting that we'd expect in
a path parameter.