Remove FlowPresenter dependency from visualising
This commit ends up doing a few things (sorry). Principally it removes
the dependency on a FlowPresenter class in the Graph*Presenter classes
and it's usage in the visualise view.

It moves both the Graph*Presenter classes to the app/presenter directory
as they seemed to be oddly placed. Given they have the suffix of
Presenter they likely belong in the app/presenters directory despite the
concept of Presenter being awfully vague and arbitrary.

It also fixes a bug where visualise smart answers didn't show a HTML title
because there is no accessing of a content_for(:title) in the layout
(instead @title is used).